Update doc/functions/openAIChat.md
I know why those happened, and can avoid them in the future!

In MLX files, you can create hyperlinks which link to lines in the same file. I tried to see whether this would translate into the markdown - turns out it hasn't. What it has done is taken those hyperlinks and just given them extra backticks with no functionality. So long as I stop trying to link between passages, this should not be an issue.

Co-authored-by: Christopher Creutzig <89011131+ccreutzi@users.noreply.github.com>