Disallow `dlarray` in prototype
Using `dlarray` in a prototype results in `dlarray` in the output, but without the dimension labels. Reason: `cast(X,'like',Y)` behaves that way for `dlarray` `Y`.

Given that we do not return higher-dimensional data anyway, it is questionable whether there is any reason one would want to allow `dlarray` in prototypes. Not handling them properly means we are better off disabling them altogether.

Without a dependency on Deep Learning Toolbox, we cannot have an automated test point for this change, unfortunately.