Remove hardcoded model support checks for OpenAI
OpenAI has started changing the support of features between model releases,
without making announcements. It is not feasible for us to keep updating the
list of which model supports what, and the end-user benefit is rather limited
in the first place. Thus, we now rely on just forwarding the server errors,
the same as we've always done for Azure and Ollama.